requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.5.14
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
distro==1.8.0
dj-database-url==2.1.0
Django==4.2
django-admin-rangefilter==0.12.1
django-import-export==3.3.7
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
drf-yasg2==1.19.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.0
fsspec==2023.10.0
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.333
langsmith==0.0.63
llama-index==0.8.66
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
oauth2client==4.1.3
odfpy==1.4.1
openai==1.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
protobuf==4.25.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
